+++ CHANGELOG.md
@@ -19,6 +19,34 @@ there is a new official change to the rules that has not been corrected, please
 
 ## [Unreleased]
 
+## [0.4.0] - 2020-09-05
+
+### Added
+
+- Added `:backward` as a valid direction (relative to the snake's orientation)
+  for `Snek.Board.Snake.step/2` and `Snek.Board.Snake.move/2`, which complements `:forward` and can
+  be used for a snake to throw a game by running into their own neck.
+
+### Fixed
+
+- In the `Snek.Ruleset.Standard` ruleset, snakes will not be eliminated by
+  colliding with snakes who are already eliminated. This was always supposed to
+  be the case, but some collisions with dead snakes were still being detected.
+- In the `Snek.Ruleset.Standard` ruleset, all alive snakes will move, even if a
+  move is not specified for every snake. Snakes without a move specified will
+  default to moving `:north` if it is the first move, otherwise continue in the
+  same direction they were already going. Previously, snakes without a move
+  specified did not move at all, which was not the correct behavior and could
+  cause invalid self-collision eliminations.
+- Corrected typespec for `Snek.Board.move_snake/3`.
+
+### Changed
+
+- Updated the documentation to clarify that snakes will not be moved if they
+  are already eliminated.
+
+## [0.3.0] - 2020-09-05
+
 ### Changed
 
 - In the `Snek.Ruleset.Standard` rules, snakes eating on their
@@ -44,7 +72,9 @@ there is a new official change to the rules that has not been corrected, please
 - A `Snek.Ruleset.Standard` module implementing the Battlesnake Standard rules.
 - A `Snek.Ruleset.Solo` module implementing the Battlesnake Solo rules.
 
-[Unreleased]: https://github.com/xtagon/snek/compare/v0.2.0...edge
+[Unreleased]: https://github.com/xtagon/snek/compare/v0.4.0...edge
+[0.4.0]: https://github.com/xtagon/snek/compare/v0.3.0...v0.4.0
+[0.3.0]: https://github.com/xtagon/snek/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/xtagon/snek/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/xtagon/snek/releases/tag/v0.1.0